// src/cometbft/cometbft-types.ts
export interface CommitResponse {
    readonly header: Header;
    readonly commit: Commit;
    readonly canonical: boolean;
}
export interface GenesisResponse {
    readonly genesisTime: ReadonlyDate;
    readonly chainId: string;
    readonly consensusParams: ConsensusParams;
    readonly validators: readonly Validator[];
    readonly appHash: Uint8Array;
    readonly appState: Record<string, unknown> | undefined;
}
export type HealthResponse = null;
export interface NumUnconfirmedTxsResponse {
    readonly total: number;
    readonly totalBytes: number;
}
export interface StatusResponse {
    readonly nodeInfo: NodeInfo;
    readonly syncInfo: SyncInfo;
    readonly validatorInfo: Validator;
}
// ...
export interface ValidatorsResponse {
    readonly blockHeight: number;
    readonly validators: readonly Validator[];
    readonly count: number;
    readonly total: number;
}
// ...
export interface BlockId {
    readonly hash: Uint8Array;
    readonly parts: {
        readonly total: number;
        readonly hash: Uint8Array;
    };
}
// ...
export interface Commit {
    readonly blockId: BlockId;
    readonly height: number;
    readonly round: number;
    readonly signatures: readonly CommitSignature[];
}
// ...
export interface Version {
    readonly block: number;
    readonly app: number;
}
export interface Header {
    readonly version: Version;
    readonly chainId: string;
    readonly height: number;
    readonly time: ReadonlyDateWithNanoseconds;
    /**
     * Block ID of the previous block. This can be `null` when the current block is height 1.
     */
    readonly lastBlockId: BlockId | null;
    /**
     * Hashes of block data.
     *
     * This is `sha256("")` for height 1 🤷‍
     */
    readonly lastCommitHash: Uint8Array;
    /**
     * This is `sha256("")` as long as there is no data 🤷‍
     */
    readonly dataHash: Uint8Array;
    readonly validatorsHash: Uint8Array;
    readonly nextValidatorsHash: Uint8Array;
    readonly consensusHash: Uint8Array;
    /**
     * This can be an empty string for height 1 and turn into "0000000000000000" later on 🤷‍
     */
    readonly appHash: Uint8Array;
    /**
     * This is `sha256("")` as long as there is no data 🤷‍
     */
    readonly lastResultsHash: Uint8Array;
    /**
     * This is `sha256("")` as long as there is no data 🤷‍
     */
    readonly evidenceHash: Uint8Array;
    readonly proposerAddress: Uint8Array;
}
export interface NodeInfo {
    readonly id: Uint8Array;
    /** IP and port */
    readonly listenAddr: string;
    readonly network: string;
    /**
     * The Tendermint version. Can be empty (see https://github.com/cosmos/cosmos-sdk/issues/7963).
     */
    readonly version: string;
    readonly channels: string;
    readonly moniker: string;
    readonly other: Map<string, string>;
    readonly protocolVersion: {
        readonly p2p: number;
        readonly block: number;
        readonly app: number;
    };
}
export interface SyncInfo {
    readonly earliestAppHash?: Uint8Array;
    readonly earliestBlockHash?: Uint8Array;
    readonly earliestBlockHeight?: number;
    readonly earliestBlockTime?: ReadonlyDate;
    readonly latestBlockHash: Uint8Array;
    readonly latestAppHash: Uint8Array;
    readonly latestBlockHeight: number;
    readonly latestBlockTime: ReadonlyDate;
    readonly catchingUp: boolean;
}
export interface Validator {
    readonly address: Uint8Array;
    readonly pubkey?: ValidatorPubkey;
    readonly votingPower: bigint;
    readonly proposerPriority?: number;
}
export interface ValidatorUpdate {
    readonly pubkey: ValidatorPubkey;
    readonly votingPower: bigint;
}
export interface ConsensusParams {
    readonly block: BlockParams | undefined;
    readonly evidence: EvidenceParams | undefined;
}
export interface BlockParams {
    readonly maxBytes: number;
    readonly maxGas: number;
}
export interface TxSizeParams {
    readonly maxBytes: number;
    readonly maxGas: number;
}
export interface BlockGossipParams {
    readonly blockPartSizeBytes: number;
}
export interface EvidenceParams {
    readonly maxAgeNumBlocks: number;
    readonly maxAgeDuration: number;
}